added swagger and controllers api routes

+++ src/city/city.service.ts
@@ -2,6 +2,7 @@ import { Injectable } from '@nestjs/common';
 import { PrismaClient } from '@prisma/client';
 import { CityCreateDto } from './dto/city-create.dto';
 import { CityUpdateDto } from './dto/city-update.dto';
+import * as bcrypt from 'bcrypt';
 
 @Injectable()
 export class CityService {
@@ -22,9 +23,12 @@ export class CityService {
         longitude: parseFloat(longitude),
       },
     });
+    const salt = await bcrypt.genSalt(5);
+    data.password = await bcrypt.hash(data.password, salt);
     return this.prisma.city.create({
       data: {
         name: data.name,
+        password: data.password,
         coordinatesId: coordinates.id,
       },
     });
@@ -33,31 +37,36 @@ export class CityService {
     const city = await this.prisma.city.findUnique({
       where: { id: id },
     });
-    const { latitude, longitude } = data;
-    const coordinates = await this.prisma.coordinates.update({
-      where: { id: city.coordinatesId },
-      data: {
-        latitude: parseFloat(latitude),
-        longitude: parseFloat(longitude),
-      },
-    });
+    if (data.latitude && data.longitude) {
+      const { latitude, longitude } = data;
+      await this.prisma.coordinates.update({
+        where: { id: city.coordinatesId },
+        data: {
+          latitude: parseFloat(latitude),
+          longitude: parseFloat(longitude),
+        },
+      });
+    }
+    if (data.password) {
+      const salt = await bcrypt.genSalt(5);
+      city.password = await bcrypt.hash(data.password, salt);
+    }
+    city.name = data.name ?? city.name;
     return this.prisma.city.update({
       where: { id: id },
       data: {
-        name: data.name,
-        coordinatesId: coordinates.id,
+        ...city,
       },
     });
   }
   async delete(id: number) {
-    const city = await this.prisma.city.findUnique({
+    const city = await this.prisma.city.delete({
       where: { id: id },
     });
+    if (!city) return null;
     await this.prisma.coordinates.delete({
       where: { id: city.coordinatesId },
     });
-    return this.prisma.city.delete({
-      where: { id: id },
-    });
+    return city;
   }
 }
